requirements.txt
aiohttp==3.10.8
aiosqlite==0.20.0
beautifulsoup4==4.12.3
discord.py==2.4.0
fuzzywuzzy==0.18.0
logger==1.4
lxml==5.3.0
mysql_connector_repackaged==0.3.1
numpy
python-dotenv==1.0.1
regex==2024.9.11
Requests==2.32.3
selenium==4.25.0
SQLAlchemy==2.0.35
chromedriver-autoinstaller==0.2.2